// src/app/context/api/time-attendance/reason.model.ts
export type IReason = {
    isUpdate: boolean;
    compId: number;
    code: number;
    reasonName: string;
    reasonNameArabic: string;
    systemCd: boolean;
    active: boolean;
    otherCd: number;
    color: string;
    rcd: string;
    createdUser: number;
    createdDt: Date;
    lastUpdatedUser: number;
    lastUpdatedDt: Date;
    exception: boolean;
    deductMissHrs: boolean;
    validationExemption: boolean;
    repcode: number;
    excuseReason: boolean;
    excuseType: number;
    leaveRequest: boolean;
    leaveCalculationDaysType: boolean;
    leaveEligibleForFullYear: boolean;
    otherLeave: boolean;
    othLvOneTime: boolean;
    othLvAnyTime: boolean;
    leaveCarryForward: boolean;
    leaveCarryForwardLimit: boolean;
    leaveCarryForwardLimitUpTo: number;
    leaveToBeImportedToPayroll: boolean;
    linkCd: string;
    leaveEligibleForFullServiceYear: boolean;
    forPayrollDeduction: boolean;
    displaySequence: number;
    appliedInProbation: boolean;
    defaultDays: number;
    firstPartYears: number;
    defaultSecondDays: number;
    isCalenderDays: boolean;
    isLeaveAttachmentRequired: boolean;
    isOffIncludedInWorkingDays: boolean;
    isLeaveHiddenInSelfService: boolean;
    isDefaultLeaveForNewEmp: boolean;
    isPayrollSpecialImport: boolean;
    showInLeaveAdjustment: boolean;
}